refactor: remove unnecessary variable assignment
In the vertical and horizontal movement methods, I was assigning the
values passed as args to the rank and file parameters, to new variables
called new_rank and new_file respectively. However this is completely
useless as I can just use the passed values instead, and it won't modify
the original values, only the passed in values.