refactor: extract repeated code into single method
Added the Movable#moves_in_direction method, to replace repeated code in
methods. I realized that I was using the same exact loop for every
method, and could just extract the code into a single method which not
only makes the code a lot simpler, it's a lot slimmer, and more readable
this way.