feat: add Player#choose_piece
Added the Player#choose_piece which will allow the player to select the
piece they want to move by entering the coordinates of the piece they
want to move. The method then calls the Player#own_movable_piece?
method, which check that the piece the player selected, is their own
piece, and is eligible to move. This prevents the player from
accidentally moving an opponent piece or somehow moving an unmovable
piece.