feat: track player's pieces in Player class
Added the @pieces instance variable to the Player class, along with the
Player#find_pieces method, to locate each piece.

Also made the Player class and observer of an instance of the Board
class, so it's own update method gets called on board updates, which
will then call the find_pieces method to update the player's piece list.