feat: implement check and checkmate
Added the Board#in_check? method to check whether the king piece of the
given color is in check according to the chess rules.

Also refactored how moving a piece was handled previously to accomodate
for implementing check and checkmate. Added Piece#safe_moves to ensure
that pieces can't make moves that would put their king in check.