feat: implement score calculation
Added a way to calculate the score for the player of a given color,
based on what pieces they have versus what pieces their opponent has,
and where their own pieces are located. This will be used to allow a
computer player to make it's moves based on the score their given moves
will get them.